+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added none output formatter to CLI commons. (Shell)
+
 ### Changed
 
 - Fix issue with duplicate variable declaration in command handlers (Shell)